refactor: utils functions for working with numbers are combined into one file
