+++ src/components/HistoryForm/utils/generateRandomData.ts
@@ -1,4 +1,4 @@
-import { getRandomIntInclusive } from '@/utils/getRandomIntInclusive';
+import { getRandomIntInclusive } from '@/utils';
 
 export function generateRandomData(rows: number): number[][] {
   const DEFAULT_MIN_VALUE = 1;
