--- src/utils/getRandomIntInclusive.ts
@@ -1,6 +1,0 @@
-export function getRandomIntInclusive(min: number, max: number) {
-  const minCeil = Math.ceil(min);
-  const maxFloor = Math.floor(max);
-
-  return Math.floor(Math.random() * (maxFloor - minCeil + 1) + minCeil);
-}
