+++ src/utils/numbers.ts
@@ -20,3 +20,10 @@ export function truncateNumber(number: number) {
 
   return number.toFixed(2);
 }
+
+export function getRandomIntInclusive(min: number, max: number) {
+  const minCeil = Math.ceil(min);
+  const maxFloor = Math.floor(max);
+
+  return Math.floor(Math.random() * (maxFloor - minCeil + 1) + minCeil);
+}
